feat(*): project setup
