+++ .prettierrc.js
@@ -9,4 +9,4 @@ module.exports = {
   tabWidth: 2,
   trailingComma: 'es5',
   useTabs: false,
-};
+}
