+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "typescript.validate.enable": true,
+  "typescript.tsdk": "./node_modules/typescript/lib",
+  "flow.enabled": false,
+  "editor.wordBasedSuggestions": true,
+  "javascript.nameSuggestions": true,
+}
