--- src/App.jsx
@@ -1,13 +1,0 @@
-// @flow
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-
-import Root from './pages';
-
-const App = () => (
-  <BrowserRouter>
-    <Root />
-  </BrowserRouter>
-};
-
-export default App;
