+++ src/App.tsx
@@ -1,0 +1,30 @@
+import * as React from 'react'
+
+import logoSvg from './logo.svg'
+
+import './App.css'
+
+class App extends React.Component {
+  render() {
+    return (
+      <div className="App">
+        <header className="App-header">
+          <img src={logoSvg} className="App-logo" alt="logo" />
+          <p>
+            Edit <code>src/App.js</code> and save to reload.
+          </p>
+          <a
+            className="App-link"
+            href="https://reactjs.org"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Learn React
+          </a>
+        </header>
+      </div>
+    )
+  }
+}
+
+export default App
