--- src/assets/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default as LogoReact } from './logo.svg';
