--- src/components/SideBar/styledComponents.js
@@ -1,6 +1,0 @@
-// @flow
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  width: 256px;
-`;
