+++ src/index.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+
+import App from './App'
+
+import * as serviceWorker from './serviceWorker'
+import './index.css'
+
+ReactDOM.render(<App />, document.getElementById('root'))
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: http://bit.ly/CRA-PWA
+serviceWorker.unregister()
