--- src/pages/LandingPage/styledComponents.js
@@ -1,9 +1,0 @@
-// @flow
-import styled from 'styled-components';
-import MUITypography from '@material-ui/core/Typography';
-
-export const Section = styled.section`
-  margin: 24px 0;
-`;
-
-export const Typography = styled(MUITypography)``;
