--- src/pages/NotFound/index.jsx
@@ -1,8 +1,0 @@
-// @flow
-import React from 'react';
-
-import * as Styled from './styledComponents';
-
-const NotFound = () => <Styled.Container>404: Page Not Found</Styled.Container>;
-
-export default NotFound;
