--- src/pages/NotFound/styledComponents.js
@@ -1,4 +1,0 @@
-// @flow
-import styled from 'styled-components';
-
-export const Container = styled.div``;
