--- src/pages/styledComponents.js
@@ -1,12 +1,0 @@
-// @flow
-import styled from 'styled-components';
-
-export const Layout = styled.div`
-  margin: 0;
-  padding: 0;
-`;
-
-export const Body = styled.div`
-  margin-top: 120px;
-  padding: 0 20%;
-`;
